SP5: create ScottPlot.OpenGL project
resolves #2399
