+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/AddPlottableExtensions.cs
