+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/GLPrograms/GLShader.cs
@@ -1,7 +1,7 @@
 ﻿using OpenTK.Graphics.OpenGL;
 using System.Diagnostics;
 
-namespace ScottPlot.WinForms.OpenGL.GLPrograms;
+namespace ScottPlot.OpenGL.GLPrograms;
 
 public class GLShader
 {
