+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/GLPrograms/IGLProgram.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ScottPlot.WinForms.OpenGL.GLPrograms;
+namespace ScottPlot.OpenGL.GLPrograms;
 
 public interface IGLProgram : IDisposable
 {
