+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/GLPrograms/MarkerOpenSquareProgram.cs
@@ -1,7 +1,7 @@
 ﻿using OpenTK.Graphics;
 using OpenTK.Graphics.OpenGL;
 
-namespace ScottPlot.WinForms.OpenGL.GLPrograms;
+namespace ScottPlot.OpenGL.GLPrograms;
 
 public class MarkerOpenSquareProgram : MarkerFillSquareProgram
 {
