+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/Plottables/ScatterGLCustom.cs
@@ -1,7 +1,7 @@
 ﻿using OpenTK.Graphics.OpenGL;
 using ScottPlot.Control;
 using ScottPlot.DataSources;
-using ScottPlot.WinForms.OpenGL.GLPrograms;
+using ScottPlot.OpenGL.GLPrograms;
 using SkiaSharp;
 using System;
