+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/nuget-readme.md
@@ -1,0 +1,3 @@
+This is a preview release of ScottPlot 5
+
+See https://scottplot.net/versions for more information
