+++ src/ScottPlot5/ScottPlot5.sln
@@ -43,6 +43,8 @@ Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "ScottPlotShared", "..\Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScottPlotSharedTests", "..\Shared\ScottPlotTests\ScottPlotSharedTests.csproj", "{5BA8842B-B3B1-42B4-A3BD-334126453C70}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScottPlot.OpenGL", "ScottPlot5 Controls\ScottPlot.OpenGL\ScottPlot.OpenGL.csproj", "{F0DE5518-62CD-4E20-A5E6-670AE7B17AF9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -109,10 +111,16 @@ Global
 		{CBF0C2BD-B5FB-47F5-95F0-B917DA97CA9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CBF0C2BD-B5FB-47F5-95F0-B917DA97CA9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CBF0C2BD-B5FB-47F5-95F0-B917DA97CA9E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{93BC39FC-E052-4B07-9149-E3120A10CFF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93BC39FC-E052-4B07-9149-E3120A10CFF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5BA8842B-B3B1-42B4-A3BD-334126453C70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5BA8842B-B3B1-42B4-A3BD-334126453C70}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5BA8842B-B3B1-42B4-A3BD-334126453C70}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5BA8842B-B3B1-42B4-A3BD-334126453C70}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F0DE5518-62CD-4E20-A5E6-670AE7B17AF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F0DE5518-62CD-4E20-A5E6-670AE7B17AF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F0DE5518-62CD-4E20-A5E6-670AE7B17AF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F0DE5518-62CD-4E20-A5E6-670AE7B17AF9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -130,6 +138,7 @@ Global
 		{09961DF4-5E0E-4779-B85E-0FD304267750} = {9C3D2BEE-76E3-4C99-9D38-07EB5CC162B9}
 		{F2C622DE-1D3D-4027-9009-5FB02428856E} = {957BAA80-422B-4231-BEC1-D03FD89425A2}
 		{CBF0C2BD-B5FB-47F5-95F0-B917DA97CA9E} = {9C3D2BEE-76E3-4C99-9D38-07EB5CC162B9}
+		{F0DE5518-62CD-4E20-A5E6-670AE7B17AF9} = {957BAA80-422B-4231-BEC1-D03FD89425A2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {FBAE082A-173A-4F13-BE31-E6C070513C9C}
